feat(weather): 扩展天气图标支持并优化UI更新机制
- 将天气图标字段从uint8_t扩展为uint16_t，支持100-9999范围的图标代码
- 实现完整的天气图标代码到Unicode字符的映射表，基于QWeather Icons字体
- 移除LVGL互斥锁保护，改用变量绑定方式更新天气相关的UI组件
- 添加四个新的UI变量(weather_icon, weather_temp, weather_text, weather_uptime)用于天气数据显示
- 在UI刷新循环中添加对天气相关控件的动态文本更新逻辑
- 优化assets数据以支持新的UI变量绑定机制

+++ main/include/weather.h
@@ -7,7 +7,7 @@
 typedef struct {
     float temperature;  // 温度，单位摄氏度
     float feelslike;    // 体感温度，单位摄氏度
-    uint8_t icon;       // 天气图标代码
+    uint16_t icon;      // 天气图标代码（支持100-9999）
     char text[32];      // 天气描述文本
     char wind_dir[8];   // 风向
     uint8_t wind_scale; // 风力等级
